+++ src/client_table_schema.py
@@ -1,35 +1,8 @@
-import sqlalchemy as db
-from sqlalchemy.ext.declarative import declarative_base
-from client_table_params import ClientTableParams
-from basic_defs import ClientType
-ClientBase = declarative_base()
+from client_table_prototype import ClientTable
 client_table_suffix = "client_history_"
 
 
 def create_client_table(table_name: str):
-
-    class ClientTable(ClientBase):
-        """
-        The main Table
-        """
-        __tablename__ = table_name
-        entry_id = db.Column(db.Integer, primary_key=True, unique=True, autoincrement=True)
-        date = db.Column(db.Date, nullable=False)
-        client_type = db.Column(db.Enum(ClientType), nullable=False)
-        tasks = db.Column(db.Unicode)
-        notes = db.Column(db.Unicode)
-
-        def __init__(self, **kwargs):
-            try:
-                params = ClientTableParams(kwargs)
-                self.entry_id = params.entry_id
-                self.date = params.date
-                self.client_type = params.client_type
-                self.tasks = params.tasks
-                self.notes = params.notes
-            except KeyError:
-                pass
-
-        def __repr__(self):
-            return f"Client_table_{self.client_id}"
-    return ClientTable
+    a = ClientTable()
+    a.set_table_name(table_name)
+    return a
