+++ src/table_schemas.py
@@ -1,9 +1,8 @@
 import sqlalchemy as db
 from sqlalchemy.ext.declarative import declarative_base
 
-from basic_defs import KnownFrom, Cities, Education, WorkType, FamilyStatus, ClientType
+from basic_defs import KnownFrom, Cities, Education, WorkType, FamilyStatus
 from main_table_params import MainTableParams
-from client_table_params import ClientTableParams
 Base = declarative_base()
 
 main_table_name = "main_table"
@@ -59,29 +58,3 @@ class MainTable(Base):
 
     def __repr__(self):
         return "MainTable"
-
-
-# class ClientTable(Base):
-#     """
-#     The main Table
-#     """
-#     __tablename__ = client_table_suffix
-#     entry_id = db.Column(db.Integer, primary_key=True, unique=True, autoincrement=True)
-#     date = db.Column(db.Date, nullable=False)
-#     client_type = db.Column(db.Enum(ClientType), nullable=False)
-#     tasks = db.Column(db.Unicode)
-#     notes = db.Column(db.Unicode)
-#
-#     def __init__(self, **kwargs):
-#         try:
-#             params = ClientTableParams(kwargs)
-#             self.entry_id = params.entry_id
-#             self.date = params.date
-#             self.client_type = params.client_type
-#             self.tasks = params.tasks
-#             self.notes = params.notes
-#         except KeyError:
-#             pass
-#
-#     def __repr__(self):
-#         return f"Client_table_{self.client_id}"
